Kenzie comments.
